requirements.txt
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.0.4
django-cors-headers==2.2.0
django-simple-history==2.0
djangorestframework==3.7.7
docker==3.2.1
docker-pycreds==0.2.2
idna==2.6
mysqlclient==1.3.12
pycparser==2.18
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
